Add userClassRepository
