+++ src/main/java/com/web/yapp/server/domain/service/ContractService.java
@@ -1,13 +1,12 @@
 package com.web.yapp.server.domain.service;
 
 import com.web.yapp.server.controller.dto.ContractRequestDto;
-import com.web.yapp.server.domain.Contract;
 import com.web.yapp.server.domain.Musician;
 import com.web.yapp.server.domain.User;
 import com.web.yapp.server.domain.repository.ContractRepository;
 import com.web.yapp.server.domain.repository.ContractTagRepository;
 import com.web.yapp.server.domain.repository.MusicianRepository;
-import com.web.yapp.server.domain.repository.UserRepository;
+import com.web.yapp.server.domain.repository.UserClassRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,7 +16,7 @@ import java.util.List;
 @Service
 public class ContractService {
     private final ContractRepository contractRepository;
-    private final UserRepository userRepository;
+    private final UserClassRepository userClassRepository;
     private final MusicianRepository musicianRepository;
     private final ContractTagRepository contractTagRepository;
 /*
@@ -30,7 +29,7 @@ public class ContractService {
     public void createContract(ContractRequestDto contractRequestDto, List<String> atmoList, List<String> themeList,
                                List<String> instruList, List<String> genreList, List<String> spclNoteList, Long userId, Long musicianId){
         Musician musician = musicianRepository.findOne(musicianId);
-        User user = userRepository.findUserById(userId);
+        User user = userClassRepository.findUserById(userId);
 //        Contract contract = contractRequestDto.toEntity(musician, user, "대기");
 //        contractRepository.save(contract);
         //태그 map 저장
